Update entry.sh (#55)
I was migrating the server from my game rig to an actual dedicated machine and wanted to copy over the autoexec.cfg to get everything up and running as seamless as possible.
To do so I added a little while loop to check for predefined variables that the entry.sh wants to force on the server.

Good contribution! Merged.